+++ resizer2/resizer2.cpp
@@ -1,293 +1,17 @@
 #define OEMRESOURCE
 #include "resizer2.h"
+#include "window_ops.h"
 
-const std::array<int, 13> systemCursors{{
-    OCR_NORMAL,
-    OCR_IBEAM,
-    OCR_WAIT,
-    OCR_CROSS,
-    OCR_UP,
-    OCR_SIZENWSE,
-    OCR_SIZENESW,
-    OCR_SIZEWE,
-    OCR_SIZENS,
-    OCR_SIZEALL,
-    OCR_NO,
-    OCR_HAND,
-    OCR_APPSTARTING,
-}};
+// systemCursors defined in window_ops.cpp
 
-Context ctx;
-HHOOK hKeyboardHook = NULL;
-HHOOK hMouseHook = NULL;
-bool modKeyPressed = false;
-bool didUseWindowsKey = false;
-bool shouldMinimize = false;
-std::chrono::steady_clock::time_point lastClickTime;
-NOTIFYICONDATA nid;
-static bool g_trayIconAdded = false;
-static bool g_hooksInstalled = false;
+// global state moved to state.cpp
 static const UINT RETRY_TIMER_ID = 1;
 static const UINT RETRY_INTERVAL_MS = 2000; // ms
-static const UINT ID_TRAY_EXIT = 0x5001;
-static HWND g_hMainWnd = NULL;
 
-static bool IsExplorerReady() {
-	return FindWindow(L"Shell_TrayWnd", NULL) != NULL;
-}
-
-static void EnsureHooksInstalled();
+extern void EnsureHooksInstalled();
 static void TryInitialize(HWND hWnd);
-static void ShowTrayMenu(HWND hWnd);
+extern void ShowTrayMenu(HWND hWnd);
 static void EnsureMessageFilter(HWND hWnd);
-
-static HWND getTopLevelParent(HWND hwnd) {
-	HWND parent = hwnd;
-	HWND tmp;
-	while ((tmp = GetParent(parent)) != NULL) {
-		parent = tmp;
-	}
-	return parent;
-}
-
-template <ContextType type>
-static void moveWindow(HWND hwnd, RECT rect, bool async = true) {
-	UINT resizeFlags = SWP_NOZORDER | SWP_NOSENDCHANGING | SWP_DEFERERASE;
-	UINT moveFlags =  resizeFlags | SWP_NOSIZE | SWP_NOREDRAW | SWP_NOCOPYBITS;
-	SetWindowPos(hwnd, NULL, rect.left, rect.top, rect.right - rect.left, rect.bottom - rect.top, (type == MOVE ? moveFlags : resizeFlags) | (async ? SWP_ASYNCWINDOWPOS : 0));
-}
-
-static void adjustRect(HWND win, RECT& rect) {
-	LONG style = GetWindowLong(win, GWL_STYLE);
-	LONG styleEx = GetWindowLong(win, GWL_EXSTYLE);
-
-	AdjustWindowRectEx(&rect, style, FALSE, styleEx);
-}
-
-
-static void snapToMonitor(HWND window, HMONITOR screen) {
-	MONITORINFO info{};
-	info.cbSize = sizeof(MONITORINFO);
-	GetMonitorInfo(screen, &info);
-
-	RECT rect = info.rcWork; // Work area, excluding taskbars
-
-	HMONITOR currentScreen = MonitorFromWindow(window, MONITOR_DEFAULTTONEAREST);
-
-	if (currentScreen == screen) {
-		return;
-	}
-
-	WINDOWPLACEMENT wp = { sizeof(WINDOWPLACEMENT) };
-	ShowWindow(window, SW_RESTORE);
-	GetWindowPlacement(window, &wp);
-
-	// Calculate the new position (center of the new screen)
-	int width = wp.rcNormalPosition.right - wp.rcNormalPosition.left;
-	int height = wp.rcNormalPosition.bottom - wp.rcNormalPosition.top;
-	
-	int x = rect.left + (rect.right - rect.left - width) / 2;
-	int y = rect.top + (rect.bottom - rect.top - height) / 2;
-	
-	MoveWindow(window, x, y, width, height, TRUE);
-	ShowWindow(window, SW_MAXIMIZE);
-}
-
-static void snapToFancyZone(HWND hWnd, HMONITOR hMon, POINT mousePos, bool maximized)
-{
-	// 1) Get the monitor's work area.
-	MONITORINFO mi{};
-	mi.cbSize = sizeof(mi);
-	if (!GetMonitorInfo(hMon, &mi)) return;
-
-	int monWidth = mi.rcWork.right - mi.rcWork.left;
-	int monHeight = mi.rcWork.bottom - mi.rcWork.top;
-
-	// 2) Compute corners, center, arcs, etc.
-
-	// Monitor top-left corner in screen coords:
-	POINT tlCorner = { mi.rcWork.left,  mi.rcWork.top };
-	// top-right corner:
-	POINT trCorner = { mi.rcWork.right, mi.rcWork.top };
-	// bottom-left corner:
-	POINT blCorner = { mi.rcWork.left,  mi.rcWork.bottom };
-	// bottom-right corner:
-	POINT brCorner = { mi.rcWork.right, mi.rcWork.bottom };
-
-	// Radius for arcs (1/3 of whichever dimension you like)
-	double radius = CORNER_RADIUS_FRACTION * min(monWidth, monHeight);
-
-	// We'll compute distance from corners:
-	auto dist = [](POINT a, POINT b) {
-		double dx = double(a.x - b.x);
-		double dy = double(a.y - b.y);
-		return sqrt(dx * dx + dy * dy);
-		};
-
-	double distTL = dist(mousePos, tlCorner);
-	double distTR = dist(mousePos, trCorner);
-	double distBL = dist(mousePos, blCorner);
-	double distBR = dist(mousePos, brCorner);
-
-
-	int centerX = mi.rcWork.left + monWidth / 2;
-	int centerY = mi.rcWork.top + monHeight / 2;
-
-	if (maximized) {
-		ShowWindow(hWnd, SW_RESTORE);
-	}
-
-	RECT newRect = mi.rcWork;
-	bool shouldMove = false;
-
-
-	// 3) Check if we're inside any corner circle
-	if (distTL <= radius) {
-		// Snap top-left quadrant
-		newRect.right -= monWidth / 2;
-		newRect.bottom -= monHeight / 2;
-		shouldMove = true;
-	}
-	else if (distTR <= radius) {
-		// Snap top-right quadrant
-		newRect.left += monWidth / 2;
-		newRect.bottom -= monHeight / 2;
-		shouldMove = true;
-	}
-	else if (distBL <= radius) {
-		// Snap bottom-left quadrant
-		newRect.right -= monWidth / 2;
-		newRect.top += monHeight / 2;
-		shouldMove = true;
-	}
-	else if (distBR <= radius) {
-		// Snap bottom-right quadrant
-		newRect.left += monWidth / 2;
-		newRect.top += monHeight / 2;
-		shouldMove = true;
-	}
-	else {
-		int oneThirdY = mi.rcWork.top + (monHeight / 3);
-		int twoThirdY = mi.rcWork.top + 2 * (monHeight / 3);
-
-		if ((mousePos.y > oneThirdY) && (mousePos.y < twoThirdY))
-		{
-			// Snap left or right half
-			if (mousePos.x < centerX) {
-				newRect.right -= monWidth / 2;
-			}
-			else {
-				newRect.left += monWidth / 2;
-			}
-			shouldMove = true;
-		}
-
-		// 6) Top/Bottom half:
-		int oneThirdX = mi.rcWork.left + (monWidth / 3);
-		int twoThirdX = mi.rcWork.left + 2 * (monWidth / 3);
-
-		if ((mousePos.x > oneThirdX) && (mousePos.x < twoThirdX))
-		{
-			if (mousePos.y < centerY) {
-				// top half
-				newRect.bottom -= monHeight / 2;
-			}
-			else {
-				// bottom half
-				newRect.top += monHeight / 2;
-			}
-			shouldMove = true;
-		}
-	}
-	
-	if (shouldMove) {
-		adjustRect(hWnd, newRect);
-		moveWindow<RESIZE>(hWnd, newRect);
-	}
-
-	// If we get here, we didn't trigger any fancy zone, so do nothing special.
-}
-
-
-static bool isWindowAllowed(HWND win) {
-	LPTSTR className = new TCHAR[256];
-	GetClassName(win, className, 256);
-
-	std::string str{ static_cast<std::string>(CT2A(className)) };
-
-	if (disallowedClasses.count(str) > 0) {
-		return false;
-	}
-
-	return true;
-}
-
-template <ResizerCursor cursor>
-void SetGlobalCursor() {
-	HCURSOR hCursor = NULL;
-	switch (cursor) {
-	case SIZEALL:
-		hCursor = LoadCursor(NULL, IDC_SIZEALL);
-		break;
-	case SIZENWSE:
-		hCursor = LoadCursor(NULL, IDC_SIZENWSE);
-		break;
-	case SIZENESW:
-		hCursor = LoadCursor(NULL, IDC_SIZENESW);
-		break;
-	default:
-		// Leave null to reset to system cursor
-		break;
-	}
-
-	if (hCursor == NULL) {
-		SystemParametersInfo(SPI_SETCURSORS, 0, nullptr, 0);
-		return;
-	}
-
-	for (auto cursor : systemCursors) {
-		HCURSOR newCursor = CopyCursor(hCursor);
-		SetSystemCursor(newCursor, cursor);
-	}
-}
-
-static void AddTrayIcon(HWND hWnd) {
-	if (g_trayIconAdded) {
-		return;
-	}
-	if (!IsExplorerReady()) {
-		return;
-	}
-
-	ZeroMemory(&nid, sizeof(nid));
-	nid.cbSize = sizeof(NOTIFYICONDATA);
-	nid.hWnd = hWnd;
-	nid.uID = TRAY_ICON_UID;
-	nid.uFlags = NIF_MESSAGE | NIF_ICON | NIF_TIP
-#ifdef NIF_SHOWTIP
-		| NIF_SHOWTIP
-#endif
-		;
-	nid.uCallbackMessage = WM_TRAYICON;  // Message for tray icon interaction
-	// Load a small shared icon for better shell behavior
-	HICON smallIcon = (HICON)LoadImage(NULL, IDI_APPLICATION, IMAGE_ICON, GetSystemMetrics(SM_CXSMICON), GetSystemMetrics(SM_CYSMICON), LR_SHARED);
-	nid.hIcon = smallIcon ? smallIcon : LoadIcon(NULL, IDI_APPLICATION);
-	lstrcpy(nid.szTip, TEXT("Resizer"));  // Tray icon tooltip
-	BOOL addOk = Shell_NotifyIcon(NIM_ADD, &nid);
-	if (addOk) {
-		g_trayIconAdded = true;
-	}
-	else {
-		// Ignore; retry via timer
-	}
-}
-
-static void RemoveTrayIcon() {
-	if (g_trayIconAdded) {
-		Shell_NotifyIcon(NIM_DELETE, &nid);
-		g_trayIconAdded = false;
-	}
-}
 
 LRESULT CALLBACK WindowProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
 	if (uMsg == WM_TASKBARCREATED) {
@@ -355,7 +79,7 @@ static BOOL CALLBACK MonitorEnumProc(HMONITOR hMonitor, HDC hdcMonitor, LPRECT l
 	return TRUE; // Continue enumerating if not found
 }
 
-static HMONITOR SysGetMonitorContainingPoint(int x, int y) {
+HMONITOR SysGetMonitorContainingPoint(int x, int y) {
 	MonitorSearchData data = { x, y, NULL };
 
 	EnumDisplayMonitors(NULL, NULL, MonitorEnumProc, (LPARAM)&data);
@@ -367,6 +91,13 @@ static HMONITOR SysGetMonitorContainingPoint(int x, int y) {
 
 	return data.hMonitor;
 }
+
+static void TryInitialize(HWND hWnd) {
+	EnsureHooksInstalled();
+	AddTrayIcon(hWnd);
+}
+
+static void EnsureMessageFilter(HWND hWnd) { }
 
 int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, PWSTR pCmdLine, int nCmdShow) {
 	const wchar_t CLASS_NAME[] = L"TrayIconWindowClass";
@@ -417,385 +148,5 @@ int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, PWSTR pCmdLine
 	if (hKeyboardHook) UnhookWindowsHookEx(hKeyboardHook);
 	if (hMouseHook) UnhookWindowsHookEx(hMouseHook);
 
-	return 0;
-}
-
-static void EnsureHooksInstalled() {
-	if (g_hooksInstalled) return;
-	HHOOK kb = SetWindowsHookEx(WH_KEYBOARD_LL, LowLevelKeyboardProc, NULL, 0);
-	HHOOK mouse = SetWindowsHookEx(WH_MOUSE_LL, LowLevelMouseProc, NULL, 0);
-	if (kb && mouse) {
-		hKeyboardHook = kb;
-		hMouseHook = mouse;
-		g_hooksInstalled = true;
-	}
-	else {
-		if (kb) UnhookWindowsHookEx(kb);
-		if (mouse) UnhookWindowsHookEx(mouse);
-	}
-}
-
-static void TryInitialize(HWND hWnd) {
-	EnsureHooksInstalled();
-	AddTrayIcon(hWnd);
-}
-
-static void ShowTrayMenu(HWND hWnd) {
-	POINT pt;
-	GetCursorPos(&pt);
-	HMENU hMenu = CreatePopupMenu();
-	if (!hMenu) return;
-	AppendMenu(hMenu, MF_STRING, ID_TRAY_EXIT, TEXT("Exit"));
-	SetForegroundWindow(hWnd);
-	UINT cmd = TrackPopupMenu(hMenu, TPM_RIGHTBUTTON | TPM_BOTTOMALIGN | TPM_LEFTALIGN | TPM_RETURNCMD, pt.x, pt.y, 0, hWnd, NULL);
-	// Ensure proper menu dismissal
-	SendMessage(hWnd, WM_NULL, 0, 0);
-	if (cmd == ID_TRAY_EXIT) {
-		PostQuitMessage(0);
-	}
-	DestroyMenu(hMenu);
-}
-
-static void EnsureMessageFilter(HWND hWnd) { }
-
-LRESULT CALLBACK LowLevelKeyboardProc(int nCode, WPARAM wParam, LPARAM lParam) {
-	KBDLLHOOKSTRUCT* pKbDllHookStruct = (KBDLLHOOKSTRUCT*)lParam;
-	if (nCode == HC_ACTION) {
-		if (modKeyPressed && (GetAsyncKeyState(VK_LWIN) & 0x8000) == 0) {
-			modKeyPressed = false;
-		}
-
-		if (pKbDllHookStruct->vkCode == VK_LWIN && pKbDllHookStruct->flags) {
-			if (wParam == WM_KEYDOWN || wParam == WM_SYSKEYDOWN) {
-				if (!modKeyPressed) {
-					didUseWindowsKey = false;
-				}
-				modKeyPressed = true;
-			}
-			else if (wParam == WM_KEYUP || wParam == WM_SYSKEYUP) {
-				modKeyPressed = false;
-				if (didUseWindowsKey) {
-					// Win + F13 does nothing, ugly workaround which stops the windows menu
-					// from appearing after correctly handled event
-
-					INPUT inputs[2] = {};
-					ZeroMemory(inputs, sizeof(inputs));
-
-					inputs[0].type = INPUT_KEYBOARD;
-					inputs[0].ki.wVk = DUMMY_KEY;
-
-					inputs[1].type = INPUT_KEYBOARD;
-					inputs[1].ki.wVk = DUMMY_KEY;
-					inputs[1].ki.dwFlags = KEYEVENTF_KEYUP;
-
-					UINT uSent = SendInput(ARRAYSIZE(inputs), inputs, sizeof(INPUT));
-				}
-			}
-		}
-	}
-	return CallNextHookEx(hKeyboardHook, nCode, wParam, lParam);
-}
-
-LRESULT CALLBACK LowLevelMouseProc(int nCode, WPARAM wParam, LPARAM lParam) {
-	MSLLHOOKSTRUCT* pMsLlHookStruct = (MSLLHOOKSTRUCT*)lParam;
-	if (nCode == HC_ACTION) {
-		if (modKeyPressed && (GetAsyncKeyState(VK_LWIN) & 0x8000) == 0) {
-			modKeyPressed = false;
-		}
-
-		switch (wParam) {
-		case WM_MOUSEWHEEL: {
-			if (modKeyPressed) {
-				didUseWindowsKey = true;
-				short delta = GET_WHEEL_DELTA_WPARAM(pMsLlHookStruct->mouseData);
-				if (delta > 0) {
-					adjustWindowOpacity(OPACITY_STEP);
-				}
-				else if (delta < 0) {
-					adjustWindowOpacity(-OPACITY_STEP);
-				}
-				return 1; // Suppress further processing
-			}
-			break;
-		}
-		case WM_LBUTTONDOWN: {
-			if (modKeyPressed && !ctx.inProgress) {
-				didUseWindowsKey = true;
-				startWindowOperation<MOVE>();
-				return 1;
-			}
-			break;
-		}
-		case WM_LBUTTONUP: {
-			if (ctx.inProgress && ctx.operationType == MOVE) {
-				didUseWindowsKey = true;
-				stopWindowOperation();
-				return 1;
-			}
-			break;
-		}
-		case WM_RBUTTONDOWN: {
-			if (modKeyPressed && !ctx.inProgress) {
-				didUseWindowsKey = true;
-				startWindowOperation<RESIZE>();
-				return 1;
-			}
-			break;
-		}
-		case WM_RBUTTONUP: {
-			if (ctx.inProgress && ctx.operationType == RESIZE) {
-				didUseWindowsKey = true;
-				stopWindowOperation();
-				return 1;
-			}
-			break;
-		}
-		// If we don't do this, weird stuff happens
-		case WM_MBUTTONDOWN: {
-			if (modKeyPressed) {
-				shouldMinimize = true;
-				didUseWindowsKey = true;
-				return 1;
-			}
-			break;
-		}
-		case WM_MBUTTONUP: {
-			if (shouldMinimize) {
-				shouldMinimize = false;
-				minimizeWindow();
-				return 1;
-			}
-			break;
-		}
-		default:
-			break;
-		}
-	}
-
-	return CallNextHookEx(hMouseHook, nCode, wParam, lParam);
-}
-
-void adjustWindowOpacity(int change) {
-	POINT pt;
-	GetCursorPos(&pt);
-	HWND hwnd = getTopLevelParent(WindowFromPoint(pt));
-
-	if ((hwnd == NULL) || !isWindowAllowed(hwnd)) {
-		return;
-	}
-
-	BYTE currentOpacity = MAX_OPACITY;
-	LONG_PTR exStyle = GetWindowLongPtr(hwnd, GWL_EXSTYLE);
-	if (exStyle & WS_EX_LAYERED) {
-		BYTE alpha;
-		DWORD flags;
-		GetLayeredWindowAttributes(hwnd, NULL, &alpha, &flags);
-		currentOpacity = alpha;
-	}
-	int newOpacity = max(MIN_OPACITY, min(MAX_OPACITY, currentOpacity + change));
-	SetWindowLongPtr(hwnd, GWL_EXSTYLE, exStyle | WS_EX_LAYERED);
-	SetLayeredWindowAttributes(hwnd, 0, (BYTE)newOpacity, LWA_ALPHA);
-}
-
-void minimizeWindow() {
-	POINT pt;
-	GetCursorPos(&pt);
-	HWND hwnd = WindowFromPoint(pt);
-	hwnd = getTopLevelParent(hwnd);
-
-	if (hwnd == NULL || !isWindowAllowed(hwnd)) {
-		return;
-	}
-
-	ShowWindow(hwnd, SW_MINIMIZE);
-}
-
-template<ContextType type>
-void startWindowOperation() {
-	ctx.inProgress = true;
-	ctx.hEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
-	ctx.operationType = type;
-
-	if (!ctx.hEvent) {
-		ctx.inProgress = false;
-		return;
-	}
-
-	GetCursorPos(&ctx.startMousePos);
-	ctx.targetWindow = getTopLevelParent(WindowFromPoint(ctx.startMousePos));
-
-	if (ctx.targetWindow == NULL || !isWindowAllowed(ctx.targetWindow)) {
-		ctx.inProgress = false;
-		CloseHandle(ctx.hEvent);
-		ctx.hEvent = NULL;
-		return;
-	}
-
-	// Double-click for toggle maximize
-	if (ctx.operationType == MOVE) {
-		SetGlobalCursor<SIZEALL>();
-		auto currentTime = std::chrono::steady_clock::now();
-		auto timeSinceLastClick = std::chrono::duration_cast<std::chrono::milliseconds>(currentTime - lastClickTime).count();
-
-		if (timeSinceLastClick <= DOUBLE_CLICK_THRESHOLD) {
-			WINDOWPLACEMENT wp = { sizeof(WINDOWPLACEMENT) };
-			GetWindowPlacement(ctx.targetWindow, &wp);
-			ShowWindow(ctx.targetWindow, wp.showCmd == SW_MAXIMIZE ? SW_RESTORE : SW_MAXIMIZE);
-
-			stopWindowOperation();
-			return;
-		}
-		lastClickTime = currentTime;
-	}
-
-	GetWindowRect(ctx.targetWindow, &ctx.startWindowRect);
-	SetForegroundWindow(ctx.targetWindow);
-
-	HANDLE hThread = CreateThread(NULL, 0, WindowOperationThreadProc, NULL, 0, NULL);
-	if (hThread != NULL) {
-		CloseHandle(hThread);
-	}
-}
-
-void stopWindowOperation() {
-	ctx.inProgress = false;
-	SetGlobalCursor<UNSET>();
-	if (ctx.hEvent != NULL) {
-		SetEvent(ctx.hEvent);
-		CloseHandle(ctx.hEvent);
-		ctx.hEvent = NULL;
-	}
-}
-
-DWORD WINAPI WindowOperationThreadProc(LPVOID lpParam) {
-	RECT currentWindowRect = ctx.startWindowRect;
-
-	if (ctx.operationType == MOVE) {
-		SetGlobalCursor<SIZEALL>();
-		WINDOWPLACEMENT wp = { sizeof(WINDOWPLACEMENT) };
-		POINT pt;
-		bool wasShiftDown = false;
-
-		while (ctx.inProgress) {
-			if (WaitForSingleObject(ctx.hEvent, 0) == WAIT_OBJECT_0) {
-				break;
-			}
-
-			GetWindowPlacement(ctx.targetWindow, &wp);
-			bool maximized = (wp.showCmd == SW_MAXIMIZE);
-
-			GetCursorPos(&pt);
-
-			int dx = pt.x - ctx.startMousePos.x;
-			int dy = pt.y - ctx.startMousePos.y;
-
-			// Check SHIFT
-			bool shiftDown = (GetAsyncKeyState(VK_SHIFT) & 0x8000) != 0;
-
-			// If SHIFT is down, we want to force-restore a maximized window
-			if (shiftDown) {
-				wasShiftDown = true;
-				// If window currently maximized, un-maximize it so we can do a "normal" snap
-				if (maximized) {
-					// Update our tracking rect to the newly-restored size
-					GetWindowRect(ctx.targetWindow, &currentWindowRect);
-				}
-
-				// SHIFT is pressed => do fancy zones
-				HMONITOR screen = SysGetMonitorContainingPoint(pt.x, pt.y);
-				snapToFancyZone(ctx.targetWindow, screen, pt, maximized);
-			}
-			else {
-				if (wasShiftDown) {
-					wasShiftDown = false;
-					RECT newRect = ctx.startWindowRect;
-					OffsetRect(&newRect, dx, dy);
-					moveWindow<RESIZE>(ctx.targetWindow, newRect);
-
-				}
-				// SHIFT is not pressed:
-				if (maximized) {
-					//
-					// Option A: Just re-snap to the monitor so user can 
-					// "drag" the window across monitors even while maximized
-					// (This is what your old code does.)
-					//
-					HMONITOR screen = SysGetMonitorContainingPoint(pt.x, pt.y);
-					snapToMonitor(ctx.targetWindow, screen);
-				}
-				else {
-					RECT rect = currentWindowRect;
-
-
-					rect.left += dx;
-					rect.right += dx;
-					rect.top += dy;
-					rect.bottom += dy;
-
-					moveWindow<MOVE>(ctx.targetWindow, rect);
-				}
-			}
-
-			 Sleep(1);
-		}
-	}
-
-	else if (ctx.operationType == RESIZE) {
-		bool isLeft = false, isTop = false;
-		int windowWidth = currentWindowRect.right - currentWindowRect.left;
-		int windowHeight = currentWindowRect.bottom - currentWindowRect.top;
-
-		// Determine which corner is nearest
-		if (ctx.startMousePos.x - currentWindowRect.left < windowWidth / 2)
-			isLeft = true;
-		if (ctx.startMousePos.y - currentWindowRect.top < windowHeight / 2)
-			isTop = true;
-
-		bool nwse = (isLeft && isTop) || (!isLeft && !isTop);
-
-		if (nwse) {
-			SetGlobalCursor<SIZENWSE>();
-		}
-		else {
-			SetGlobalCursor<SIZENESW>();
-		}
-
-		// Restore window if maximized
-		WINDOWPLACEMENT wp = { sizeof(WINDOWPLACEMENT) };
-		GetWindowPlacement(ctx.targetWindow, &wp);
-		if (wp.showCmd == SW_MAXIMIZE) {
-			ShowWindow(ctx.targetWindow, SW_RESTORE);
-		}
-
-		while (ctx.inProgress) {
-			if (WaitForSingleObject(ctx.hEvent, 0) == WAIT_OBJECT_0) {
-				break;
-			}
-			POINT pt;
-			GetCursorPos(&pt);
-			int dx = pt.x - ctx.startMousePos.x;
-			int dy = pt.y - ctx.startMousePos.y;
-
-			RECT newRect = currentWindowRect;
-			if (isLeft) {
-				newRect.left += dx;
-			}
-			else {
-				newRect.right += dx;
-			}
-			if (isTop) {
-				newRect.top += dy;
-			}
-			else {
-				newRect.bottom += dy;
-			}
-
-			moveWindow<RESIZE>(ctx.targetWindow, newRect, false);
-
-			Sleep(1);
-		}
-	}
-
-	ctx.inProgress = false;
 	return 0;
 }
