+++ resizer2/resizer2.h
@@ -1,11 +1,15 @@
 #pragma once
 
+#ifndef OEMRESOURCE
+#define OEMRESOURCE
+#endif
+
 #include <windows.h>
 #include <iostream>
 #include <array>
 #include <chrono>
 #include <unordered_set>
-#include <atlbase.h>
+#include <vector>
 #include <string>
 #include <shellapi.h>
 #include <WtsApi32.h>
@@ -16,6 +20,7 @@
 
 #define TRAY_ICON_UID 1001
 #define WM_TRAYICON (WM_APP + 1)
+#define ID_TRAY_EXIT 0x5001
 
 // Config
 constexpr int DOUBLE_CLICK_THRESHOLD = 300; // ms
@@ -25,17 +30,17 @@ constexpr BYTE OPACITY_STEP = 26; // Around 10% of 255
 constexpr int DUMMY_KEY = VK_F13; // Any key that doesn't do anything when pressed together with the Windows key
 constexpr double CORNER_RADIUS_FRACTION = 0.33; // 1/3 of monitor dimension
 
-const std::unordered_set<std::string> disallowedClasses{
-    "ControlCenterWindow",
-    "TopLevelWindowForOverflowXamlIsland",
-    "XamlExplorerHostIslandWindow",
-    "Windows.UI.Core.CoreWindow",
-    "Shell_InputSwitchDismissOverlay",
-    "Shell_InputSwitchTopLevelWindow",
-    "Shell_SecondaryTrayWnd",
-    "Shell_TrayWnd",
-    "Progman",
-	"WorkerW",
+const std::unordered_set<std::wstring> disallowedClasses{
+	L"ControlCenterWindow",
+	L"TopLevelWindowForOverflowXamlIsland",
+	L"XamlExplorerHostIslandWindow",
+	L"Windows.UI.Core.CoreWindow",
+	L"Shell_InputSwitchDismissOverlay",
+	L"Shell_InputSwitchTopLevelWindow",
+	L"Shell_SecondaryTrayWnd",
+	L"Shell_TrayWnd",
+	L"Progman",
+	L"WorkerW",
 };
 
 enum ContextType {
@@ -75,11 +80,16 @@ extern bool didUseWindowsKey;
 extern bool shouldMinimize;
 extern std::chrono::steady_clock::time_point lastClickTime;
 extern NOTIFYICONDATA nid;
+extern bool g_trayIconAdded;
+extern bool g_hooksInstalled;
+extern HWND g_hMainWnd;
 
 LRESULT CALLBACK LowLevelKeyboardProc(int nCode, WPARAM wParam, LPARAM lParam);
 LRESULT CALLBACK LowLevelMouseProc(int nCode, WPARAM wParam, LPARAM lParam);
 DWORD WINAPI WindowOperationThreadProc(LPVOID lpParam);
-UINT WM_TASKBARCREATED;
+extern UINT WM_TASKBARCREATED;
+void EnsureHooksInstalled();
+void ShowTrayMenu(HWND hWnd);
 
 void adjustWindowOpacity(int change);
 void minimizeWindow();
@@ -88,6 +98,7 @@ void startWindowOperation();
 void stopWindowOperation();
 HWND getTopLevelParent(HWND hwnd);
 void snapToMonitor(HWND window, HMONITOR screen);
+void snapToFancyZone(HWND hWnd, HMONITOR hMon, POINT mousePos, bool maximized);
 template <ResizerCursor cursor>
 void SetGlobalCursor();
 void AddTrayIcon(HWND hWnd);
